+++ framework/Source/CPTAnimation.h
@@ -120,4 +120,9 @@ CPTAnimationCurve;
 -(void)removeAllAnimationOperations;
 /// @}
 
+/// @name Retrieving Animation Operations
+/// @{
+-(CPTAnimationOperation *)operationWithIdentifier:(id<NSCopying, NSObject>)identifier;
+/// @}
+
 @end